requirements.txt
Django==3.2.6
djangorestframework
pillow
dj_database_url
django-cors-headers
django-markdownx
django-url-framework
psycopg2